+++ src/GlobalSettings.cpp
@@ -2,15 +2,15 @@
 
 namespace StoryTime {
 
-unsigned int GlobalSettings::WINDOWWIDTH = 10;
-unsigned int GlobalSettings::WINDOWHEIGHT = 10;
+size_t GlobalSettings::WINDOWWIDTH = 10;
+size_t GlobalSettings::WINDOWHEIGHT = 10;
 Markup GlobalSettings::currentSettings = Markup();
 sf::Font GlobalSettings::DEFAULTFONT = sf::Font();
 std::filesystem::path GlobalSettings::ROOTDIR = "";
 std::filesystem::path GlobalSettings::SAVEDIR = "";
 bool GlobalSettings::INITIALISED = false;
-unsigned int GlobalSettings::CHARSIZE = 10;
-unsigned int GlobalSettings::LINESPACE = 10;
+size_t GlobalSettings::CHARSIZE = 10;
+size_t GlobalSettings::LINESPACE = 10;
 
 void GlobalSettings::initialise()
 {
@@ -37,24 +37,24 @@ void GlobalSettings::initialise()
   INITIALISED = true;
 }
 
-void GlobalSettings::setCharSize(unsigned int newSize)
+void GlobalSettings::setCharSize(size_t newSize)
 {
   CHARSIZE = newSize;
   LINESPACE = newSize * 2;
 }
 
-void GlobalSettings::setLineSpacing(unsigned int newSize)
+void GlobalSettings::setLineSpacing(size_t newSize)
 {
   CHARSIZE = newSize / 2;
   LINESPACE = newSize;
 }
 
-int GlobalSettings::getCharSize()
+size_t GlobalSettings::getCharSize()
 {
   return CHARSIZE;
 }
 
-int GlobalSettings::getLineSpacing()
+size_t GlobalSettings::getLineSpacing()
 {
   return LINESPACE;
 }
