+++ src/OpeningScreen.cpp
@@ -12,12 +12,12 @@ OpeningScreen::OpeningScreen() : GameSlice()
 
   _titleText.setFont(_titleFont);
   _titleText.setString("StoryTime");
-  _titleText.setCharacterSize(GlobalSettings::WINDOWHEIGHT / 10);
+  _titleText.setCharacterSize(static_cast<uint>(GlobalSettings::WINDOWHEIGHT / 10));
   _titleText.setFillColor(sf::Color::White);
   _titleText.setStyle(sf::Text::Bold | sf::Text::Underlined);
   _titleTextRect = _titleText.getLocalBounds();
   _titleText.setOrigin(_titleTextRect.left + _titleTextRect.width / 2.0f, _titleTextRect.top + _titleTextRect.height / 2.0f);
-  _titleText.setPosition(sf::Vector2f(GlobalSettings::WINDOWWIDTH / 2.0f, GlobalSettings::WINDOWHEIGHT / 2.0f));
+  _titleText.setPosition(sf::Vector2f(static_cast<float>(GlobalSettings::WINDOWWIDTH) / 2.0f, static_cast<float>(GlobalSettings::WINDOWHEIGHT) / 2.0f));
   redraw();
 }
