+++ src/SelectStory.cpp
@@ -15,18 +15,18 @@ SelectStory::SelectStory() : GameSlice()
   _windowBorder.setPosition(0, 0);
 
   sf::Vector2f orgSize = (sf::Vector2f)_borderTexture.getSize();
-  _windowBorder.setScale(GlobalSettings::WINDOWWIDTH / orgSize.x, GlobalSettings::WINDOWHEIGHT / orgSize.y);
+  _windowBorder.setScale(static_cast<float>(GlobalSettings::WINDOWWIDTH) / orgSize.x, static_cast<float>(GlobalSettings::WINDOWHEIGHT) / orgSize.y);
 
   int count = 1;
   for (auto& story : std::filesystem::directory_iterator("stories"))
   {
     if (std::filesystem::is_directory(story))
     {
-      sf::Text storyText(story.path().stem().string(), GlobalSettings::DEFAULTFONT, GlobalSettings::WINDOWHEIGHT / GlobalSettings::getCharSize() * 1.5);
+      sf::Text storyText(story.path().stem().string(), GlobalSettings::DEFAULTFONT, static_cast<uint>(static_cast<float>(GlobalSettings::WINDOWHEIGHT) / static_cast<float>(GlobalSettings::getCharSize()) * 1.5f));
       storyText.setFillColor(sf::Color::White);
       sf::FloatRect textRect = storyText.getLocalBounds();
       storyText.setOrigin(textRect.left + textRect.width / 2.0f, textRect.top + textRect.height / 2.0f);
-      storyText.setPosition(sf::Vector2f(GlobalSettings::WINDOWWIDTH / 2.0f, GlobalSettings::WINDOWHEIGHT / 4.0f * (1 + count * 0.4)));
+      storyText.setPosition(sf::Vector2f(static_cast<float>(GlobalSettings::WINDOWWIDTH) / 2.0f, static_cast<float>(GlobalSettings::WINDOWHEIGHT) / 4.0f * static_cast<float>(1 + count * 0.4)));
       _storyChoices.push_back(storyText);
       count++;
     }
@@ -34,12 +34,12 @@ SelectStory::SelectStory() : GameSlice()
 
   _titleText.setFont(GlobalSettings::DEFAULTFONT);
   _titleText.setString("Select A Story To Play");
-  _titleText.setCharacterSize(GlobalSettings::WINDOWHEIGHT / GlobalSettings::getCharSize() * 3);
+  _titleText.setCharacterSize(static_cast<uint>(GlobalSettings::WINDOWHEIGHT / GlobalSettings::getCharSize() * 3));
   _titleText.setFillColor(sf::Color::White);
   _titleText.setStyle(sf::Text::Bold | sf::Text::Underlined);
   sf::FloatRect textRect = _titleText.getLocalBounds();
   _titleText.setOrigin(textRect.left + textRect.width / 2.0f, textRect.top + textRect.height / 2.0f);
-  _titleText.setPosition(sf::Vector2f(GlobalSettings::WINDOWWIDTH / 2.0f, GlobalSettings::WINDOWHEIGHT / 4.0f));
+  _titleText.setPosition(sf::Vector2f(static_cast<float>(GlobalSettings::WINDOWWIDTH) / 2.0f, static_cast<float>(GlobalSettings::WINDOWHEIGHT) / 4.0f));
 
   _backgroundColour = sf::Color::Black;
   _changeSlice = false;
